requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0
requests>=2.28.0

# LangChain framework for multi-agent systems
langchain>=0.1.0

# For data handling
pandas>=2.0.0
numpy>=1.24.0

# GUI
streamlit>=1.28.0

# File processing
PyPDF2>=3.0.0
python-docx>=0.8.11
pathlib2>=2.3.0